fix: refactor page capacity calc into benches subdir
