+++ core/src/benches/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(all(test, feature = "calculate-page-capacity"))]
+mod page_size;
