+++ core/src/graph/page.rs
@@ -636,51 +636,3 @@ mod test {
 		assert_eq!(graph_page, graph_page2);
 	}
 }
-
-#[cfg(all(test, feature = "calculate-page-capacity"))]
-mod page_capacity {
-	use dsnp_graph_config::ALL_CONNECTION_TYPES;
-
-	use super::*;
-	use crate::tests::helpers::*;
-	use std::{collections::hash_map::HashMap, path::PathBuf};
-
-	#[test]
-	fn calculate_page_capacities() {
-		let mut capacity_map: HashMap<ConnectionType, usize> = HashMap::new();
-
-		for c in ALL_CONNECTION_TYPES {
-			let mut result_vec: Vec<usize> = Vec::new();
-			for _ in 0..1000 {
-				result_vec.push(benchmark_page_capacity(c).0);
-			}
-			result_vec.sort();
-			capacity_map.insert(c, *result_vec.first().unwrap());
-		}
-
-		let code = format!(
-			"use dsnp_graph_config::{{ConnectionType, ConnectionType::Follow, ConnectionType::Friendship, PrivacyType::Public, PrivacyType::Private}};
-use lazy_static::lazy_static;
-use std::collections::hash_map::*;
-
-lazy_static! {{
-	pub static ref PAGE_CAPACITIY_MAP: HashMap<ConnectionType, usize> = {{
-		let m = HashMap::from({:?});
-		m
-	}};
-}}
-",
-			capacity_map.iter().collect::<Vec::<(&ConnectionType, &usize)>>()
-		);
-
-		let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-		path.push("src/graph/page_capacities.rs");
-
-		let result = std::fs::write(path, code);
-		if result.is_err() {
-			println!("Error: {:?}", result);
-		}
-
-		assert!(true);
-	}
-}
