+++ core/src/lib.rs
@@ -6,3 +6,6 @@ mod graph;
 mod tests;
 mod types;
 mod util;
+
+#[cfg(all(test, feature = "calculate-page-capacity"))]
+mod benches;
